frontend_api/requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
redis
dj-database-url